requirements.txt
torch>=2.1.0
diffusers>=0.19.0
transformers>=4.40.0
accelerate>=0.22.0
safetensors>=0.3.1
gradio>=3.46.0
python-dotenv>=1.0.0
google-genai>=0.1.0
psutil>=5.9.5
Pillow>=10.0.0
